requirements.txt
# Core dependencies
llama-parse>=0.4.0
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-llms-gemini>=0.1.0

# HTTP Client for OpenRouter
httpx>=0.27.0

# Data processing
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
tenacity>=8.2.0

# Optional: MongoDB integration
pymongo>=4.6.0
motor>=3.3.0
